safari-600.7-branch/LayoutTests/platform/ios-sim/fast/forms/ValidityState-typeMismatch-week-expected.txt

This test aims to check for typeMismatch flag with type=week input fields

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "" is a correct valid week string.
PASS "0001-W01" is a correct valid week string.
PASS "1583-W01" is a correct valid week string.
PASS "9999-W52" is a correct valid week string.
PASS "275760-W37" is a correct valid week string.
PASS "2009-W01" is a correct valid week string.
PASS "2004-W53" is a correct valid week string.
PASS "2003-W52" is a correct valid week string.
PASS "1992-W53" is a correct valid week string.
FAIL " 2009-W01 " is an incorrect valid week string.
FAIL "2009W01" is an incorrect valid week string.
FAIL "2009-w01" is an incorrect valid week string.
FAIL "2009-W1" is an incorrect valid week string.
FAIL "2009-W001" is an incorrect valid week string.
FAIL "a" is an incorrect valid week string.
FAIL "-1-W09" is an incorrect valid week string.
FAIL "0000-W52" is an incorrect valid week string.
FAIL "2147483648-W01" is an incorrect valid week string.
FAIL "275760-W38" is an incorrect valid week string.
FAIL "2009-W00" is an incorrect valid week string.
FAIL "2009-W-1" is an incorrect valid week string.
FAIL "2004-W54" is an incorrect valid week string.
FAIL "2003-W53" is an incorrect valid week string.
FAIL "1992-W54" is an incorrect valid week string.
FAIL "2009/09" is an incorrect valid week string.
FAIL "200909" is an incorrect valid week string.
FAIL "2009-Wxx" is an incorrect valid week string.
FAIL "2009" is an incorrect valid week string.
PASS "invalid" is a correct valid week string when disabled.
PASS successfullyParsed is true

TEST COMPLETE
